Create shared crypto package for password encryption
Move password encryption/decryption to pkg/crypto with random salt format.
Remove legacy username-as-salt encryption code from both collector and
server. All passwords are now encrypted with cryptographically random
salts using AES-256-GCM.

Update documentation to reflect the new encryption approach and remove
references to manual password encryption. Users should use the MCP server
API to manage connection passwords.

Closes #6
