+++ pkg/go.mod
@@ -2,4 +2,7 @@ module github.com/pgedge/ai-workbench/pkg
 
 go 1.24.0
 
-require gopkg.in/yaml.v3 v3.0.1
+require (
+	golang.org/x/crypto v0.47.0
+	gopkg.in/yaml.v3 v3.0.1
+)
